+++ src/main/java/gp/cnusambe/payload/response/OssLicenseListResponse.java
@@ -3,18 +3,17 @@ package gp.cnusambe.payload.response;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import gp.cnusambe.dto.OssLicenseDto;
+import gp.cnusambe.dto.PageInfoDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import org.springframework.data.domain.Page;
 
-import java.util.HashMap;
 import java.util.List;
 
 @Getter
 @AllArgsConstructor
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class OssLicenseListResponse {
-    private MetaResponse meta;
+    private PageInfoDto meta;
     private List<OssLicenseDto> OssLicense;
 
 }
